Refactor translation cleanup tests: split into modular test files for validation, patterns, edge cases, and nested translations.
